+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Change Log
 ==========
 
+next
+----
+
+- Added ActiveQuery decorator.
+
 0.1, July 25, 2014
 ------------------